Create intitial Linked-list data structure
Created a Linked-list factory that will create a linked-list data
structure.
Created a Node factory that will create an object with data and a link
to the next object.
Created a linked-list method prepend that will add a data at the start
of the list.
Created a linked-list method head() that will return the head of the
list.