Add 'tail', 'append', 'size', & 'at' methods
Created 'tail' method to show the last node of the linked-list data
structure.
Created 'append' method to add a data at the end of the linked-list data
structure.
Created 'size' method to show the size of the data structure.
Created 'at' method to show the node at the specific data structure
index.