Add pop, contains, find, toString methods
Created pop method to delete the last node of the Linked-list data
structure.
Created constains method to check if the Linked-list data structure
contains the data.
Created find method to find the index of the data in the Linked-list.
Created toString method to show print all the data of the Linked-list as
a string.